% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schedule.R
\name{load_schedule}
\alias{load_schedule}
\title{Pull schedule data for a given year}
\usage{
load_schedule(years)
}
\arguments{
\item{year}{numeric}
}
\value{
list
}
\description{
Create a list of data frames -- one for each year specified in the function
}
\examples{
load_schedule(years = c(2020,2021))
}
